refactor(folder): extract model & attributes
