+++ src/modules/file/file.repository.ts
@@ -2,8 +2,6 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { File, FileAttributes, FileOptions } from './file.domain';
 import sequelize, { FindOptions, Op } from 'sequelize';
-import { FolderModel } from '../folder/folder.repository';
-
 import {
   AllowNull,
   BelongsTo,
@@ -19,8 +17,10 @@ import {
 } from 'sequelize-typescript';
 import { UserModel } from '../user/user.model';
 import { User } from '../user/user.domain';
-import { Folder, FolderAttributes } from '../folder/folder.domain';
+import { Folder } from '../folder/folder.domain';
 import { Pagination } from '../../lib/pagination';
+import { FolderModel } from '../folder/folder.model';
+import { FolderAttributes } from '../folder/folder.attributes';
 
 @Table({
   underscored: true,
