+++ src/modules/folder/folder.attributes.ts
@@ -1,0 +1,15 @@
+export interface FolderAttributes {
+    id: number;
+    parentId: number;
+    parent?: any;
+    name: string;
+    bucket: string;
+    userId: number;
+    uuid: string;
+    user?: any;
+    encryptVersion: '03-aes';
+    deleted: boolean;
+    deletedAt: Date;
+    createdAt: Date;
+    updatedAt: Date;
+}
