+++ src/modules/folder/folder.module.ts
@@ -1,6 +1,7 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { SequelizeModule } from '@nestjs/sequelize';
-import { FolderModel, SequelizeFolderRepository } from './folder.repository';
+import { SequelizeFolderRepository } from './folder.repository';
+import { FolderModel } from './folder.model';
 import { FolderUseCases } from './folder.usecase';
 import { FileModule } from '../file/file.module';
 import { CryptoModule } from '../../externals/crypto/crypto.module';
