+++ src/modules/share/share.controller.ts
@@ -34,6 +34,7 @@ import { File, FileAttributes } from '../file/file.domain';
 import { ShareDto } from './dto/share.dto';
 import { Folder } from '../folder/folder.domain';
 import { ReferralKey, User } from '../user/user.domain';
+import { Share } from './share.domain';
 
 @ApiTags('Share')
 @Controller('storage/share')
