+++ src/modules/share/share.domain.ts
@@ -1,6 +1,8 @@
 import { aes } from '@internxt/lib';
 
 import { File, FileAttributes } from '../file/file.domain';
+import { Folder } from '../folder/folder.domain';
+import { FolderAttributes } from '../folder/folder.attributes';
 import { UserAttributes } from '../user/user.attributes';
 import { ShareDto } from './dto/share.dto';
