+++ src/modules/share/share.repository.ts
@@ -19,7 +19,9 @@ import { FileModel } from '../file/file.repository';
 import { User } from '../user/user.domain';
 import { UserAttributes } from '../user/user.attributes';
 import { UserModel } from '../user/user.model';
+import { FolderModel } from '../folder/folder.model';
 import { Folder } from '../folder/folder.domain';
+import { FolderAttributes } from '../folder/folder.attributes';
 import { Pagination } from '../../lib/pagination';
 import { Op } from 'sequelize';
