+++ src/shared-workspace/shared-workspace.repository.ts
@@ -12,7 +12,7 @@ import {
   Table,
 } from 'sequelize-typescript';
 import { UserAttributes } from '../modules/user/user.attributes';
-import { UserModel } from '../modules/user/user.repository';
+import { UserModel } from '../modules/user/user.model';
 
 interface FriendInvitationAttributes {
   id: number;
